backend/requirements.txt
nltk
fastapi
text2emotion
dotenv
requests
uvicorn
mysql.connector
bcrypt
pydantic
sqlalchemy
psycopg2-binary
streamlit
pyjwt
passlib[bcrypt]
python-decouple
emoji==1.7.0
mistralai
plotly
pandas
mysql-connector-python